tech(orga): move constant modules to app/constants
