+++ orga/app/components/statistics/tag-level.gjs
@@ -2,7 +2,7 @@ import PixTag from '@1024pix/pix-ui/components/pix-tag';
 import Component from '@glimmer/component';
 import { t } from 'ember-intl';
 
-import { ADVANCED_LEVEL, INDEPENDENT_LEVEL, NOVICE_LEVEL } from '../../helpers/levels-info';
+import { ADVANCED_LEVEL, INDEPENDENT_LEVEL, NOVICE_LEVEL } from '../../constants/levels-info';
 
 export default class TagLevel extends Component {
   get category() {
