+++ orga/app/utils/params-validator.ts
@@ -1,4 +1,4 @@
-import { CERTIFICABILITY_TYPES } from '../helpers/certificability-types';
+import { CERTIFICABILITY_TYPES } from '../constants/certificability-types';
 
 export function validateCertificabilityParams(params: { certificability?: string | string[] }): void {
   if (params.certificability) {
