+++ orga/tests/integration/components/campaign/header/tabs-test.gjs
@@ -4,7 +4,7 @@ import { click } from '@ember/test-helpers';
 import { t } from 'ember-intl/test-support';
 import Tabs from 'pix-orga/components/campaign/header/tabs';
 import ENV from 'pix-orga/config/environment';
-import { EVENT_NAME } from 'pix-orga/helpers/metrics-event-name';
+import { EVENT_NAME } from 'pix-orga/constants/metrics-event-name';
 import { module, test } from 'qunit';
 import sinon from 'sinon';
