+++ orga/tests/integration/components/catalogue/course-modal-test.gjs
@@ -2,7 +2,7 @@ import { render, within } from '@1024pix/ember-testing-library';
 import { click } from '@ember/test-helpers';
 import { t } from 'ember-intl/test-support';
 import CourseModal from 'pix-orga/components/catalogue/course-modal';
-import { EVENT_NAME } from 'pix-orga/helpers/metrics-event-name';
+import { EVENT_NAME } from 'pix-orga/constants/metrics-event-name';
 import { module, test } from 'qunit';
 import sinon from 'sinon';
